pyproject.toml
[project]
name = "shop-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "langchain-google-genai>=3.2.0",
    "langgraph>=1.0.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "supabase>=2.24.0",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
]
